Improve handling of future and past on index page
We basically split the list of posts into 3 parts:

- the next meetup
- meetups _after_ the next one
- meetups in the past

This is necessary to allow having more than one meetup planned and on
the website (cf. #14).

It also makes use of some newer Jekyll filters which make the code much
more functionally-flavored and readable.

I also change the "Help us [plan the next meetup]" link to point to
Github issues (instead of being a mailto link).